+++ mammoth/__init__.py
@@ -155,7 +155,7 @@ from mammoth.models.workspaces import (
 )
 from mammoth.view import View, ViewExport
 
-__version__ = "0.7.8"
+__version__ = "0.7.9"
 __all__ = [
     # Client
     "MammothClient",